// features/summary/types/summary.ts
export type MonthlySummary = {
  adiantamento: number;
  month: number;
  renda_extra_amt: number;
  restante_adiantamento: number;
  restante_renda_extra: number;
  restante_salario: number;
  salario: number;
  total_expense: number;
  total_geral_disponivel: number;
  total_income: number;
  year: number;
};